% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_optim.R
\name{class_optim}
\alias{class_optim}
\title{Creation of dummies variables and others variables}
\usage{
class_optim(i, f.obj, f.con, f.dir, f.rhs)
}
\arguments{
\item{i}{Semester to be considered (from 1 to 3)}

\item{f.obj}{Objective function}

\item{f.con}{Constraints matrix}

\item{f.dir}{Sign of equality of constraints}

\item{f.rhs}{Right hand side of constraints}
}
\value{
Final choice of classes of user
}
\description{
This function is used to perform the integer optimization.
It can be used within the ShinyApp settings.
}
\author{
Matteo Gross, Alessia Di Pietro, Martina Celic, Ana Gabriela Garcia, Laura Lo Priore
}
