% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_text_timetable.R
\name{display_text_timetable}
\alias{display_text_timetable}
\title{Creation of a appealing textual timetable of classes}
\usage{
display_text_timetable(i, choice)
}
\arguments{
\item{i}{Semester to be considered (from 1 to 3)}

\item{choice}{Result of the function class_optim}
}
\value{
Visually appealing textual timetable object
}
\description{
This function is used to display a table timetable.
}
\author{
Group 2 composed of Matteo Gross, Alessia Di Pietro, Martina Celic, Ana Gabriela Garcia, Laura Lo Priore
}
