% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hectimetables-package.R
\docType{package}
\name{hectimetables-package}
\alias{hectimetables}
\alias{hectimetables-package}
\title{hectimetables: HEC timetable creation - functions for ShinyApp}
\description{
This packages helps to create a timetable for HEC students. It contains many functions that can be retrieved when creating the ShinyApp.
}
\keyword{internal}
