% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runDemo.R
\name{runDemo}
\alias{runDemo}
\title{Demonstration of shiny app}
\usage{
runDemo()
}
\value{
Open the app for Master BA students in 2020-2022
}
\description{
This function is used to demonstrate how the shiny app works.
}
\examples{
\dontrun{runDemo()}
}
\author{
Group 2 composed of Matteo Gross, Alessia Di Pietro, Martina Celic, Ana Gabriela Garcia, Laura Lo Priore
}
