% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_optim.R
\name{class_optim}
\alias{class_optim}
\title{Integer optimization of timetable}
\usage{
class_optim(s, f.obj, f.con, f.dir, core, elective, moment, class)
}
\arguments{
\item{s}{Semester to be considered (numerical)}

\item{f.obj}{Objective function (to be minimized)}

\item{f.con}{Constraints matrix}

\item{f.dir}{Sign of equality of constraints (">=", "<=", ...) (numerical vectors)}

\item{core}{Core credits chosen by the user by semester (numerical value)}

\item{elective}{Elective credits chosen by the user (numerical value)}

\item{moment}{Moments chosen by the user (binary vector)}

\item{class}{Classes chosen by the user (binary vector)}
}
\value{
Final choice of classes of user (binary vector)
}
\description{
This function is used to perform the integer optimization using lpSolve::lp. 




For more information, please refer to the lpSolve package in R.
}
\author{
Group 2 composed of Matteo Gross, Alessia Di Pietro, Martina Celic, Ana Gabriela Garcia, Laura Lo Priore
}
