% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_optim.R
\name{class_optim}
\alias{class_optim}
\title{Creation of dummies variables and others variables}
\usage{
class_optim(s, f.obj, f.con, f.dir, credits, core, elective, moment, class)
}
\arguments{
\item{s}{Semester to be considered (from 1 to 3)}

\item{f.obj}{Objective function}

\item{f.con}{Constraints matrix}

\item{f.dir}{Sign of equality of constraints}

\item{credits}{Total credits choosen by the user}

\item{core}{Core credits choosen by the user#'}

\item{elective}{Elective credits choosen by the user}

\item{moment}{Moments choosen by the user}

\item{class}{Moments choosen by the user}
}
\value{
Final choice of classes of user
}
\description{
This function is used to perform the integer optimization.
It can be used within the ShinyApp settings.
}
\author{
Matteo Gross, Alessia Di Pietro, Martina Celic, Ana Gabriela Garcia, Laura Lo Priore
}
