% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_visual_timetable.R
\name{display_visual_timetable}
\alias{display_visual_timetable}
\title{Creation of a appealing visual timetable of classes}
\usage{
display_visual_timetable(i, choice)
}
\arguments{
\item{i}{Semester to be considered}

\item{choice}{Result of the function class_optim}
}
\value{
Visually appealing timetable ggplot2 object
}
\description{
This function is used to display a visual timetable using ggplot2.
}
\author{
Group 2 composed of Matteo Gross, Alessia Di Pietro, Martina Celic, Ana Gabriela Garcia, Laura Lo Priore
}
