% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_creation.R
\name{dummy_creation}
\alias{dummy_creation}
\title{Creation of dummies variables and others variables}
\usage{
dummy_creation()
}
\value{
Returns many variables in the Global Environement
}
\description{
This function is used to generate necessary dummy variables.
It can be used to optimise the timetable schedule of HEC students.
}
\author{
Matteo Gross, Alessia Di Pietro, Martina Celic, Ana Gabriela Garcia, Laura Lo Priore
}
